+++ src/device/bot.ts
@@ -630,9 +630,9 @@ export class Bot {
         this.debugLog(`${this.device.deviceType}: ${this.accessory.displayName} Bot Mode: ${this.botMode}`);
         switchbot
           .discover({ model: 'H', quick: true, id: this.device.bleMac })
-          .then((device_list: { press: (arg0: { id: string | undefined }) => any }[]) => {
+          .then(async (device_list: { press: (arg0: { id: string | undefined }) => any }[]) => {
             this.infoLog(`${this.device.deviceType}: ${this.accessory.displayName} On: ${this.On}`);
-            return device_list[0].press({ id: this.device.bleMac });
+            return await device_list[0].press({ id: this.device.bleMac });
           })
           .then(() => {
             this.debugLog(`${this.device.deviceType}: ${this.accessory.displayName} Done.`);
@@ -659,12 +659,12 @@ export class Bot {
           .then(async (device_list: any) => {
             this.infoLog(`${this.device.deviceType}: ${this.accessory.displayName} On: ${this.On}`);
             return await this.retry({
-              max: await this.maxRetry(),
-              fn: () => {
+              max: this.maxRetry(),
+              fn: async () => {
                 if (this.On) {
-                  return device_list[0].turnOn({ id: this.device.bleMac });
+                  return await device_list[0].turnOn({ id: this.device.bleMac });
                 } else {
-                  return device_list[0].turnOff({ id: this.device.bleMac });
+                  return await device_list[0].turnOff({ id: this.device.bleMac });
                 }
               },
             });
@@ -1173,25 +1173,23 @@ export class Bot {
   }
 
   async retry({ max, fn }: { max: number; fn: { (): any; (): Promise<any> } }): Promise<null> {
-    return fn().catch(async (err: any) => {
+    return fn().catch(async (e: any) => {
       if (max === 0) {
-        throw err;
+        throw e;
       }
-      this.infoLog(err);
+      this.infoLog(e);
       this.infoLog(`${this.device.deviceType}: ${this.accessory.displayName} Retrying`);
       await sleep(1000);
       return this.retry({ max: max - 1, fn });
     });
   }
 
-  async maxRetry(): Promise<number> {
-    let maxRetry: number;
+  maxRetry(): number {
     if (this.device.maxRetry) {
-      maxRetry = this.device.maxRetry;
+      return this.device.maxRetry;
     } else {
-      maxRetry = 5;
+      return 5;
     }
-    return maxRetry;
   }
 
   async PressOrSwitch(device: device & devicesConfig): Promise<void> {